requirements.txt
# StimuPop - Production Dependencies
# Python 3.8+

# Web Framework
streamlit>=1.24.0

# PowerPoint Generation
python-pptx>=0.6.21

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.0

# Image Processing
Pillow>=9.0.0

# Configuration
PyYAML>=6.0
